+++ composer.json
@@ -8,6 +8,7 @@
     ],
 
     "autoload": {
-        "classmap": ["entities"]
-    }
+      "classmap": ["src/entities"]
+     }
+
 }
