+++ index.php
@@ -1,6 +1,16 @@
 <?php
 
-require "vendor/autoload.php";
+use src\entities\Bicycle;
+use src\entities\Boat;
+use src\entities\Bus;
+use src\entities\Car;
+use src\entities\Motorcycle;
+use src\entities\Plane;
+
+require_once "vendor/autoload.php";
+
+
+
 
 
 $bicycle = new Bicycle("shimano", "29", "amarillo", 1);
