+++ src/entities/boat.php
@@ -1,5 +1,6 @@
 <?php
 
+namespace src\entities;
 require_once "Vehicle.php";
 
 class Boat extends Vehicle
